Extract institution events
