+++ common/src/main/scala/hmda/messages/institution/InstitutionEvents.scala
@@ -1,0 +1,12 @@
+package hmda.messages.institution
+
+import hmda.model.institution.Institution
+
+object InstitutionEvents {
+
+  sealed trait InstitutionEvent
+  case class InstitutionCreated(i: Institution) extends InstitutionEvent
+  case class InstitutionModified(i: Institution) extends InstitutionEvent
+  case class InstitutionDeleted(LEI: String) extends InstitutionEvent
+  case class InstitutionNotExists(LEI: String) extends InstitutionEvent
+}
