+++ hmda/src/main/resources/persistence.conf
@@ -24,10 +24,10 @@ akka {
       "hmda.persistence.institution.InstitutionPersistence$GetInstitution" = institution-commands
       "hmda.persistence.institution.InstitutionPersistence$DeleteInstitution" = institution-commands
 
-      "hmda.persistence.institution.InstitutionPersistence$InstitutionCreated" = institution-events
-      "hmda.persistence.institution.InstitutionPersistence$InstitutionModified" = institution-events
-      "hmda.persistence.institution.InstitutionPersistence$InstitutionDeleted" = institution-events
-      "hmda.persistence.institution.InstitutionPersistence$InstitutionNotExists" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionCreated" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionModified" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionDeleted" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionNotExists" = institution-events
     }
   }
 }
