+++ hmda/src/main/scala/hmda/persistence/institution/InstitutionPersistence.scala
@@ -5,6 +5,7 @@ import akka.cluster.sharding.typed.scaladsl.EntityTypeKey
 import akka.persistence.typed.scaladsl.{Effect, PersistentBehaviors}
 import akka.persistence.typed.scaladsl.PersistentBehaviors.CommandHandler
 import com.typesafe.config.ConfigFactory
+import hmda.messages.institution.InstitutionEvents._
 import hmda.model.institution.Institution
 
 object InstitutionPersistence {
@@ -13,26 +14,16 @@ object InstitutionPersistence {
 
   sealed trait InstitutionCommand
 
-  sealed trait InstitutionEvent
-
   case class CreateInstitution(i: Institution,
                                replyTo: ActorRef[InstitutionCreated])
       extends InstitutionCommand
-
-  case class InstitutionCreated(i: Institution) extends InstitutionEvent
 
   case class ModifyInstitution(i: Institution,
                                replyTo: ActorRef[InstitutionEvent])
       extends InstitutionCommand
 
-  case class InstitutionModified(i: Institution) extends InstitutionEvent
-
   case class DeleteInstitution(LEI: String, replyTo: ActorRef[InstitutionEvent])
       extends InstitutionCommand
-
-  case class InstitutionDeleted(LEI: String) extends InstitutionEvent
-
-  case class InstitutionNotExists(LEI: String) extends InstitutionEvent
 
   case class GetInstitution(replyTo: ActorRef[Option[Institution]])
       extends InstitutionCommand
