+++ hmda/src/main/scala/hmda/persistence/serialization/institution/InstitutionCommandsProtobufConverter.scala
@@ -4,6 +4,7 @@ import hmda.persistence.institution.InstitutionPersistence._
 import hmda.persistence.serialization.institution.commands._
 import InstitutionProtobufConverter._
 import akka.actor.typed.ActorRefResolver
+import hmda.messages.institution.InstitutionEvents.InstitutionEvent
 
 object InstitutionCommandsProtobufConverter {
